requirements.txt
appnope
backcall
beautifulsoup4
certifi
chardet
cymem
Cython
cytoolz
decorator
dill
en-core-web-sm
google
idna
ipython
ipython-genutils
jedi
lxml
msgpack-numpy
msgpack-python
murmurhash
nltk
numpy
parso
pathlib
pexpect
pickleshare
Pillow
plac
preshed
prompt-toolkit
ptyprocess
PyAutoGUI
pycodestyle
Pygments
PyMsgBox
pyobjc
pyobjc-core
pyobjc-framework-Accounts
pyobjc-framework-AddressBook
pyobjc-framework-AppleScriptKit
pyobjc-framework-AppleScriptObjC
pyobjc-framework-ApplicationServices
pyobjc-framework-Automator
pyobjc-framework-AVFoundation
pyobjc-framework-AVKit
pyobjc-framework-BusinessChat
pyobjc-framework-CalendarStore
pyobjc-framework-CFNetwork
pyobjc-framework-CloudKit
pyobjc-framework-Cocoa
pyobjc-framework-Collaboration
pyobjc-framework-ColorSync
pyobjc-framework-Contacts
pyobjc-framework-ContactsUI
pyobjc-framework-CoreBluetooth
pyobjc-framework-CoreData
pyobjc-framework-CoreLocation
pyobjc-framework-CoreML
pyobjc-framework-CoreServices
pyobjc-framework-CoreSpotlight
pyobjc-framework-CoreText
pyobjc-framework-CoreWLAN
pyobjc-framework-CryptoTokenKit
pyobjc-framework-DictionaryServices
pyobjc-framework-DiskArbitration
pyobjc-framework-EventKit
pyobjc-framework-ExceptionHandling
pyobjc-framework-ExternalAccessory
pyobjc-framework-FinderSync
pyobjc-framework-FSEvents
pyobjc-framework-GameCenter
pyobjc-framework-GameController
pyobjc-framework-GameKit
pyobjc-framework-GameplayKit
pyobjc-framework-ImageCaptureCore
pyobjc-framework-IMServicePlugIn
pyobjc-framework-InputMethodKit
pyobjc-framework-InstallerPlugins
pyobjc-framework-InstantMessage
pyobjc-framework-Intents
pyobjc-framework-IOSurface
pyobjc-framework-iTunesLibrary
pyobjc-framework-LatentSemanticMapping
pyobjc-framework-LaunchServices
pyobjc-framework-libdispatch
pyobjc-framework-LocalAuthentication
pyobjc-framework-MapKit
pyobjc-framework-MediaAccessibility
pyobjc-framework-MediaLibrary
pyobjc-framework-MediaPlayer
pyobjc-framework-ModelIO
pyobjc-framework-MultipeerConnectivity
pyobjc-framework-NetFS
pyobjc-framework-NetworkExtension
pyobjc-framework-NotificationCenter
pyobjc-framework-OpenDirectory
pyobjc-framework-Photos
pyobjc-framework-PhotosUI
pyobjc-framework-PreferencePanes
pyobjc-framework-PubSub
pyobjc-framework-QTKit
pyobjc-framework-Quartz
pyobjc-framework-SafariServices
pyobjc-framework-SceneKit
pyobjc-framework-ScreenSaver
pyobjc-framework-ScriptingBridge
pyobjc-framework-SearchKit
pyobjc-framework-Security
pyobjc-framework-SecurityFoundation
pyobjc-framework-SecurityInterface
pyobjc-framework-ServiceManagement
pyobjc-framework-Social
pyobjc-framework-SpriteKit
pyobjc-framework-StoreKit
pyobjc-framework-SyncServices
pyobjc-framework-SystemConfiguration
pyobjc-framework-Vision
pyobjc-framework-WebKit
PyScreeze
pytesseract
PyTweening
regex
requests
simplegeneric
six
spacy
termcolor
tesseract-ocr
thinc
toolz
tqdm
traitlets
ujson
urllib3
wcwidth
wikipedia
wrapt
